--- src/main/java/me/xhsun/guildwars2wrapper/requests/v2/achievements/daily/Daily.java
@@ -1,44 +1,0 @@
-package me.xhsun.guildwars2wrapper.requests.v2.achievements.daily;
-
-import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
-import me.xhsun.guildwars2wrapper.model.v2.achievement.DailyAchievement;
-import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
-import me.xhsun.guildwars2wrapper.requests.Request;
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
-public class Daily extends Request {
-
-	protected Daily(GuildWars2API gw2API) {
-		super(gw2API);
-	}
-
-	//ASYNC
-	/**
-	 * For more info on achievements daily API go <a href="https://wiki.guildwars2.com/wiki/API:2/achievements/daily">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see DailyAchievement daily achievement info
-	 */
-	public void getCurrentDailyAchievements(Callback<DailyAchievement> callback) throws NullPointerException {
-		gw2API.getCurrentDailyAchievements().enqueue(callback);
-	}
-
-	/**
-	 * For more info on next achievements daily API go <a href="https://wiki.guildwars2.com/wiki/API:2/achievements/daily/tomorrow">here</a><br/>
-	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
-	 *
-	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
-	 * @throws GuildWars2Exception  invalid API key
-	 * @throws NullPointerException if given {@link Callback} is empty
-	 * @see DailyAchievement daily achievement info
-	 */
-	public void getNextDailyAchievements(Callback<DailyAchievement> callback) throws GuildWars2Exception, NullPointerException {
-		gw2API.getNextDailyAchievements().enqueue(callback);
-	}
-
-	//SYNC
-}
